requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

langchain-pinecone==0.1.0


# WARNING(pigar): the following duplicate requirements are for import name: bs4
beautifulsoup4==4.12.3
bs4==0.0.2
BS4-for-aiogram==0.0.0.5
p01.build==0.8.1
web-parser==0.1.0
# WARNING(pigar): the following duplicate requirements are for import name: langchain
langchain==0.1.16

# WARNING(pigar): the following duplicate requirements are for import name: langchain_community

langchain-community==0.0.34
# WARNING(pigar): the following duplicate requirements are for import name: langchain_core

langchain-core==0.1.46
# WARNING(pigar): the following duplicate requirements are for import name: langchain_openai

langchain-openai==0.1.4
# WARNING(pigar): the following duplicate requirements are for import name: langchain_text_splitters

langchain-text-splitters==0.0.1
# WARNING(pigar): the following duplicate requirements are for import name: pinecone
pinecone-client==3.2.2

